[refactor]#20 파일 이동 및 변수 변경(StackBottomTab -> BottomStack)
